add 后端初始化 add 用户登录、注册 add 发布帖子，删除帖子 add 帖子收藏，点赞 add 帖子推荐、帖子搜索；搜索用户 add 查询发布、点赞帖子、收藏帖子

+++ src/main/java/com/hjj/xiantao/model/domain/Post.java
@@ -50,12 +50,12 @@ public class Post implements Serializable {
     /**
      * 点赞数
      */
-    private Integer thumbNum;
+    private Long thumbNum;
 
     /**
      * 收藏数
      */
-    private Integer favourNum;
+    private Long favourNum;
 
     /**
      * 创建用户 id
